refactor: rename MainLayout -> ChatLayout
